PRC-171 : API - swap all sync endpoint PUT / POST requests, refactor into better packages, update the swagger document
