--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/resource/ContactAddressSyncController.kt
@@ -1,91 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppscontactsapi.resource
-
-import io.swagger.v3.oas.annotations.Operation
-import io.swagger.v3.oas.annotations.Parameter
-import io.swagger.v3.oas.annotations.tags.Tag
-import jakarta.validation.Valid
-import org.springframework.http.HttpStatus
-import org.springframework.http.MediaType
-import org.springframework.security.access.prepost.PreAuthorize
-import org.springframework.web.bind.annotation.DeleteMapping
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.PutMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.ResponseBody
-import org.springframework.web.bind.annotation.ResponseStatus
-import org.springframework.web.bind.annotation.RestController
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.CreateContactAddressRequest
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.UpdateContactAddressRequest
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.service.SyncService
-
-@Tag(name = "Sync endpoints")
-@RestController
-@RequestMapping(value = ["/sync"], produces = [MediaType.APPLICATION_JSON_VALUE])
-class ContactAddressSyncController(
-  val syncService: SyncService,
-) {
-  @GetMapping(path = ["/contact-address/{contactAddressId}"], produces = [MediaType.APPLICATION_JSON_VALUE])
-  @ResponseStatus(HttpStatus.OK)
-  @Operation(
-    summary = "Returns the data for a contact address by contactAddressId",
-    description = """
-      Requires role: ROLE_CONTACTS_MIGRATION.
-      Used to get the details for one contact address.
-      """,
-  )
-  @PreAuthorize("hasAnyRole('CONTACTS_MIGRATION')")
-  fun getContactAddressById(
-    @Parameter(description = "The internal ID for a contact address.", required = true)
-    @PathVariable contactAddressId: Long,
-  ) = syncService.getContactAddressById(contactAddressId)
-
-  @DeleteMapping(path = ["/contact-address/{contactAddressId}"], produces = [MediaType.APPLICATION_JSON_VALUE])
-  @ResponseStatus(HttpStatus.OK)
-  @Operation(
-    summary = "Deletes one contact address by internal ID",
-    description = """
-      Requires role: ROLE_CONTACTS_MIGRATION.
-      Used to delete a contact address.
-      """,
-  )
-  @PreAuthorize("hasAnyRole('CONTACTS_MIGRATION')")
-  fun deleteContactAddressById(
-    @Parameter(description = "The internal ID for the contact address.", required = true)
-    @PathVariable contactAddressId: Long,
-  ) = syncService.deleteContactAddressById(contactAddressId)
-
-  @PutMapping(path = ["/contact-address"], produces = [MediaType.APPLICATION_JSON_VALUE])
-  @ResponseStatus(HttpStatus.OK)
-  @ResponseBody
-  @Operation(
-    summary = "Creates a new contact address",
-    description = """
-      Requires role: ROLE_CONTACTS_MIGRATION.
-      Used to create a contact address and associate it with a contact.
-      """,
-  )
-  @PreAuthorize("hasAnyRole('CONTACTS_MIGRATION')")
-  fun createContactAddress(
-    @Valid @RequestBody createContactAddressRequest: CreateContactAddressRequest,
-  ) = syncService.createContactAddress(createContactAddressRequest)
-
-  @PostMapping(path = ["/contact-address/{contactAddressId}"], produces = [MediaType.APPLICATION_JSON_VALUE])
-  @ResponseStatus(HttpStatus.OK)
-  @ResponseBody
-  @Operation(
-    summary = "Updates a contact address with new or extra detail",
-    description = """
-      Requires role: ROLE_CONTACTS_MIGRATION.
-      Used to update a contact address.
-      """,
-  )
-  @PreAuthorize("hasAnyRole('CONTACTS_MIGRATION')")
-  fun updateContactAddress(
-    @Parameter(description = "The internal ID for the contact address.", required = true)
-    @PathVariable contactAddressId: Long,
-    @Valid @RequestBody updateContactAddressRequest: UpdateContactAddressRequest,
-  ) = syncService.updateContactAddress(contactAddressId, updateContactAddressRequest)
-}
